LayoutTests/fast/dom/HTMLSelectElement/select-selectedIndex-noAnchorIndex-crash.html

<!DOCTYPE html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>

<select id="multiselect" multiple="multiple" style="margin-top: 20px;">
    <option>item a</option>
    <option>item b</option>
</select>

<script>

var multiselect = document.getElementById('multiselect');

if (window.eventSender) {
    eventSender.mouseMoveTo(multiselect.offsetLeft + 10, multiselect.offsetTop - 10);
    eventSender.mouseDown();
    eventSender.mouseMoveTo(multiselect.offsetLeft + 10, multiselect.offsetTop + 10);
    eventSender.mouseUp();
    shouldBe("multiselect.selectedIndex", "-1");
} else {
    document.write("To manually test, press left button down above the select, then move the mouse down over the select.");
}
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
